% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice.R
\name{read_south_seaice}
\alias{read_south_seaice}
\alias{read_south_seaice}
\alias{read_north_seaice}
\title{Sea ice concentration data}
\usage{
read_south_seaice(date, latest = TRUE, ...)

read_north_seaice(date, latest = TRUE, ...)
}
\arguments{
\item{date}{input date}

\item{latest}{return the latest day available, otherwise the earliest available}

\item{...}{ignored}
}
\value{
RasterLayer
}
\description{
`read_south_seaice` for the southern hemisphere
`read_north_seaice`  for the northern hemisphere
}
\examples{
\dontrun{
if (seaice_files_available()) {
read_south_seaice(latest = FALSE)
}
}
# w <- sp::spTransform(wrld_simpl, projection(read_north_seaice()))
#image(read_north_seaice("2017-01-01"), col = viridis::viridis(100), asp = 1)
#plot(w, add = T)
#image(read_north_seaice(latest = T), col = viridis::viridis(100), asp = 1)
#plot(w, add = T)
}
