% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaice.R
\name{read_south_seaice}
\alias{read_south_seaice}
\alias{read_south_seaice}
\alias{read_north_seaice}
\title{Sea ice concentration data}
\usage{
read_south_seaice(date, latest = TRUE, ...)

read_north_seaice(date, latest = TRUE, ...)
}
\arguments{
\item{date}{input date}

\item{latest}{return the latest day available, otherwise the earliest available}

\item{...}{ignored}
}
\value{
RasterLayer
}
\description{
`read_south_seaice` will read daily sea ice concentration data for the southern hemisphere
`read_north_seaice` will read daily sea ice concentration data for the northern hemisphere
}
\examples{
\dontrun{
if (seaice_files_available()) {
read_south_seaice(latest = FALSE)
}
}
}
